psalm magic ✨

+++ src/Discounts/Handlers/SwitchesCategoryDiscountHandler.php
@@ -13,9 +13,9 @@ class SwitchesCategoryDiscountHandler implements DiscountHandler
         $items = array_filter($order->getItems(), fn (Item $item) => $item->getCategory() === '1');
 
         foreach ($items as $item) {
-            $freeItemsCount = (int) floor((int) $item->getQuantity() / 6);
+            $freeItemsCount = floor($item->getQuantity() / 6);
 
-            $itemDiscount = (float) $item->getPrice() * $freeItemsCount;
+            $itemDiscount = $item->getPrice() * $freeItemsCount;
 
             $order->applyDiscount(new Discount(
                 "sixth_switch_product_bought_{$item->getId()}",
